+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed:
 - Update JournalCard in infinite scroll automatically on change, e.g. after navigating back
+- Clear query
 
 ## [0.8.227] - 2023-01-07
 ### Changed: